Update iota.yml
I'm the developer behind iota-nodes.net. I've created an automated nodelist using the nelson peer discovery protocol. The total amount of public nodes is larger than the provided amount. Not all nodes are using the nelson protocol and therefore the real number can't be determined. Since the nelson protocol has a big share the provided number can be used as a good approximation. The current number of public nodes using the nelson protocol can always be found  on https://iota-nodes.net/statistics and is refreshed hourly.